+++ tests/resources/components/functions/declared.blade.php
@@ -1,0 +1,9 @@
+<?php
+\Stillat\Dagger\component()->props(['title'])->trimOutput();
+
+function toUpper($value) {
+  return mb_strtoupper($value);
+}
+?>
+
+{{ toUpper($title) }}