+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Adds a `#cache` compiler attribute, which may be used to cache the results of any Dagger component
 - Bumps the minimum Laravel version to `11.23`, for `Cache::flexible` support
 - Improves compilation of custom functions declared within a component's template
+- Reduces overall memory utilization
+- Simplifies serialized output of dynamic and circular components
 
 ## [v1.0.6](https://github.com/Stillat/dagger/compare/v1.0.5...v1.0.6) - 2025-01-31