+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [v1.1.0](https://github.com/Stillat/dagger/compare/v1.0.6...v1.1.0) - 2025-02-08
 
 - Adds a new "Compile Time Rendering" system, which can render components at compile time and inline the static output
 - Adds compiler support for circular component references, like nested comment threads